Remove '?' prefix in FeedOption::build_url
Repeated inclusion of "?" to identify the query string will cause
any parameters to which "?" is appended to be treated as "param?".
This leads to the parameter being ignored by the Reddit API as it
cannot be parsed correctly.

For now, let's remove the "?" appending in FeedOption::build_url
and instead include it in every URL passed to build_url.